==> scripts/02-download_images.ipynb <==
import os
import pandas as pd

import requests
from PIL import Image
from io import BytesIO

os.makedirs('../PH_IMAGES_DIR', exist_ok=True)
data = pd.read_csv('../data/PH_CO2_geotiff_cropped_filtered.csv')

def get_static_images(filename, lat, lon):
    base_url = 'https://maps.googleapis.com/maps/api/staticmap?'
    center = (lat, lon)
    zoom = 13
    size = (400, 400)
    maptype = 'satellite'
    key = '<API KEY>' #BLOCK THIS OUT BEFORE YOU UPLOAD TO GITHUB

    gsm_url = base_url+'center='+str(center[0])+','+str(center[1])+'&zoom='+str(zoom)+'&size='+str(size[0])+'x'+str(size[1])+'&maptype='+maptype+'&key='+key
    response = requests.get(gsm_url)
    img = Image.open(BytesIO(response.content))
    img.convert("RGB").save("../PH_IMAGES_DIR/"+filename)

downloaded = os.listdir('../PH_IMAGES_DIR')

data['filename'] = str(data['y']) + '_' + str(data['x']) + '.jpg'
for i in range(len(data)):
    data['filename'][i] = str(data['y'][i]) + '_' + str(data['x'][i]) + '.jpg'

data = data.set_index('filename').drop(downloaded).reset_index()

#downloads the images

for i in range(len(data)):
    get_static_images(str(data['y'][i])+'_'+str(data['x'][i])+'.jpg', data['y'][i], data['x'][i])

data.rename(columns = {'y':'lat', 'x':'lon', 'z':'emission'}, inplace=True)
data

data.to_csv(('../data/data.csv'), index=False)
