==> scripts/03-append-nightlights.ipynb <==
import pandas as pd
import numpy as np
import os
import geoio

data = pd.read_csv('../data/data.csv')

df_clusters = data
df_clusters.head()

#img = geoio.GeoImage('../../data/ntl.tif')
img = geoio.GeoImage('../data/odiac2019_1km_excl_intl_1812.tif') ###THIS SHOULD BE THE NIGHTLIGHTS FILE, MAKE SURE TO CHANGE

# pass lon then lat
xPixel, yPixel = img.proj_to_raster(18.612500, 121.087500)

xPixel, yPixel

im_array = np.squeeze(img.get_data())
im_array.shape

im_array[int(yPixel),int(xPixel)] # this is the nightlight value at the given coordinate

import math

def create_space(lat, lon):
    # these are pulled from the paper to make the 10km^2 area
    return lat - (180/math.pi)*(5000/6378137), lon - (180/math.pi)*(5000/6378137)/math.cos(lat), \
            lat + (180/math.pi)*(5000/6378137), lon + (180/math.pi)*(5000/6378137)/math.cos(lat)

cluster_nightlights = []
for i,r in df_clusters.iterrows():
    min_lat, min_lon, max_lat, max_lon = create_space(r.lat, r.lon)
    xminPixel, yminPixel = img.proj_to_raster(min_lon, min_lat)
    xmaxPixel, ymaxPixel = img.proj_to_raster(max_lon, max_lat)
    
    xminPixel, xmaxPixel = min(xminPixel, xmaxPixel), max(xminPixel, xmaxPixel)
    yminPixel, ymaxPixel = min(yminPixel, ymaxPixel), max(yminPixel, ymaxPixel)
    
    xminPixel, yminPixel, xmaxPixel, ymaxPixel = int(xminPixel), int(yminPixel), int(xmaxPixel), int(ymaxPixel)
    cluster_nightlights.append(im_array[yminPixel:ymaxPixel,xminPixel:xmaxPixel].mean())

df_clusters['nightlights'] = cluster_nightlights
df_clusters.head()

df_clusters.to_csv ('../data/PH_data.csv', index = False, header=True)

test = df_clusters

from sklearn.mixture import GaussianMixture as GMM
X = test['nightlights'].values.reshape(-1,1)
gmm = GMM(n_components=3).fit(X)
labels = gmm.predict(test['nightlights'].values.reshape(-1,1))

(labels==0).mean(), (labels==1).mean(), (labels==2).mean()

# these are the cutoff for the labels identified by the Guassian Mixture Model
label0_max = test['nightlights'][labels==0].max()
label1_max = test['nightlights'][labels==1].max()
label2_max = test['nightlights'][labels==2].max()

label0_max, label1_max, label2_max

# I am going to hand reassign these to have better representation among all three classes
# these are not ideal distributions obviously but the model should still be able to learn
# something like a quantile cut might work better and be less arbitrary, but for reproducability 
# purposes I'll stick to the GMM-based approach
label0_max = 0.05
label1_max = 5
label2_max = 70

def query_df(df, lower, upper):
    return df[((lower <= df['nightlights']) & (df['nightlights'] < upper))]

print(len(query_df(test, 0, label0_max)) / len(test))
print(len(query_df(test, label0_max, label1_max)) / len(test))
print(len(query_df(test, label1_max, label2_max)) / len(test))

def create_nightlights_bin(df, cutoffs):
    assert len(cutoffs) >= 2, print('need at least 2 bins')
    cutoffs = sorted(cutoffs, reverse=True)
    labels = list(range(len(cutoffs)))[::-1]
    df['nightlights_bin'] = len(cutoffs)
    for cutoff, label in zip(cutoffs, labels):
        df['nightlights_bin'].loc[df['nightlights'] <= cutoff] = label

df_download = test.copy()
create_nightlights_bin(df_download, cutoffs=[label0_max, label1_max, label2_max])

# these should match above
(df_download['nightlights_bin']==0).mean(), (df_download['nightlights_bin']==1).mean(), (df_download['nightlights_bin']==2).mean()

df_download['filename'] = data['filename']
df_download

df_download.to_csv('../data/image_download_locs.csv')
