==> scripts/04-train_cnn.ipynb <==
import os
import shutil
import numpy as np
import pandas as pd
from tqdm.notebook import tqdm

BASE_DIR = '../scripts/data'
IMAGE_DIR2 = '../../scripts/PH_IMAGES_DIR'
RANDOM_SEED = 7 # for reproducibility

# these relate to training the CNN to predict nightlights
CNN_TRAIN_IMAGE_DIR = BASE_DIR
CNN_SAVE_DIR = os.path.join(BASE_DIR, 'models')
os.makedirs(CNN_SAVE_DIR, exist_ok=True)

# Preprocess

df_download = pd.read_csv('../data/image_download_locs.csv')

df_download.head()

# the distribution
(df_download['nightlights_bin']==1).mean(), (df_download['nightlights_bin']==2).mean(), (df_download['nightlights_bin']==0).mean()


Creates train/valid folders

The model will use these directly.

# we need this renaming because images can have the same name when they belong to different clusters
# this makes unique image names, even though some are really the same image
#we probably dpont need this rite - kirsten 
#df_download['image_renamed'] = df_download['lat'].astype(str) + '_' + df_download['lon'].astype(str) + '_' + df_download['filename']


df_download.head()

pic_list = df_download['filename'].values.tolist()
to_pick = int(0.8*len(pic_list)); to_pick

np.random.seed(RANDOM_SEED)
inds = np.arange(len(pic_list))
train_ind = np.random.choice(np.arange(len(pic_list)), to_pick, replace=False)
valid_ind = np.delete(inds, train_ind)

pic_list = np.array(pic_list)
train_im = pic_list[train_ind]
valid_im = pic_list[valid_ind]

os.makedirs(os.path.join(IMAGE_DIR2, 'train'), exist_ok=False)
os.makedirs(os.path.join(IMAGE_DIR2, 'valid'), exist_ok=False)

labels = ['1', '2', '0']
for l in labels:
    os.makedirs(os.path.join(IMAGE_DIR2, 'train', l), exist_ok=False)
    os.makedirs(os.path.join(IMAGE_DIR2, 'valid', l), exist_ok=False)

t = df_download.iloc[train_ind]
v = df_download.iloc[valid_ind]

t

len(t), len(v)

# copies images to their respective folders
# YES this does mean an image can be in train and valid, but that's ok
# because we do not really care about accuracy on nightlights
print('copying train images')
for im_name, im_renamed, l in tqdm(zip(t['filename'], t['filename'], t['nightlights_bin']), total=len(t)):
    shutil.copy(os.path.join(IMAGE_DIR2, im_name), os.path.join(IMAGE_DIR2, 'train', str(l), im_renamed))

    #shutil.copy(os.path.join(IMAGE_DIR2, im_name), os.path.join(IMAGE_DIR2, 'train', str(l), im_renamed))
    # i removed the str(l) bec we only have one image per cluster/point anyways -kirsten 
    
print('copying valid images')
for im_name, im_renamed, l in tqdm(zip(v['filename'], v['filename'], v['nightlights_bin']), total=len(v)):
    shutil.copy(os.path.join(IMAGE_DIR2, im_name), os.path.join(IMAGE_DIR2, 'valid', str(l), im_renamed))

# shows count distribution in train folder, make sure this matches above
counts = []
for l in ['1', '2', '0']:
    counts.append(len(os.listdir(os.path.join(IMAGE_DIR2, 'train', l))))
print(counts)
print([c/sum(counts) for c in counts])
print(sum(counts))

# shows count distribution in valid folder
counts = []
for l in ['1', '2', '0']:
    counts.append(len(os.listdir(os.path.join(IMAGE_DIR2, 'valid', l))))
print(counts)
print([c/sum(counts) for c in counts])
print(sum(counts))

# Train Model
Heavily adapted from the PyTorch CNN training tutorial.

import torch
import torch.nn as nn
import torch.optim as optim
import numpy as np
import torchvision
from torchvision import datasets, models, transforms
import matplotlib.pyplot as plt
import time
import os
import copy

# Top level data directory.

data_dir = IMAGE_DIR2
# Models to choose from [resnet, alexnet, vgg, squeezenet, densenet, inception]
model_name = "vgg"

# Number of classes in the dataset
num_classes = 3

# Batch size for training (change depending on how much memory you have)
batch_size = 8

# Number of epochs to train for
num_epochs = 5

# Flag for feature extracting. When False, we finetune the whole model,
#   when True we only update the reshaped layer params
feature_extract = True

def initialize_model(model_name, num_classes, feature_extract, use_pretrained=True):
    # Initialize these variables which will be set in this if statement. Each of these
    #   variables is model specific.
    model_ft = None
    input_size = 0

    if model_name == "vgg":
        """ VGG11_bn
        """
        model_ft = models.vgg11_bn(pretrained=use_pretrained)
        set_parameter_requires_grad(model_ft, feature_extract)
        num_ftrs = model_ft.classifier[6].in_features
        model_ft.classifier[6] = nn.Linear(num_ftrs,num_classes)
        input_size = 224
    else:
        print("Invalid model name, exiting...")
        exit()

    return model_ft, input_size

def set_parameter_requires_grad(model, feature_extracting):
    if feature_extracting:
        for param in model.parameters():
            param.requires_grad = False

# Initialize the model for this run
model_ft, input_size = initialize_model(model_name, num_classes, feature_extract, use_pretrained=True)

# Print the model we just instantiated
model_ft

# you can modify the classifier part of the model by doing this
# model_ft.classifier = nn.Sequential(
#     nn.Linear(in_features=25088, out_features=4096, bias=True),
#     nn.ReLU(inplace=True),
#     nn.Dropout(p=0.5),
#     nn.Linear(in_features=4096, out_features=256, bias=True),
#     nn.ReLU(inplace=True),
#     nn.Dropout(p=0.5),
#     nn.Linear(in_features=256, out_features=3, bias=True),
# )

# Data augmentation and normalization for training
# Just normalization for validation
data_transforms = {
    'train': transforms.Compose([
        transforms.RandomResizedCrop(input_size),
        transforms.RandomHorizontalFlip(),
        transforms.ToTensor(),
        transforms.Normalize([0.485, 0.456, 0.406], [0.229, 0.224, 0.225])
    ]),
    'valid': transforms.Compose([
        transforms.Resize(input_size),
        transforms.CenterCrop(input_size),
        transforms.ToTensor(),
        transforms.Normalize([0.485, 0.456, 0.406], [0.229, 0.224, 0.225])
    ]),
}
print("Initializing Datasets and Dataloaders...")

# Create training and validation datasets
image_datasets = {x: datasets.ImageFolder(os.path.join(data_dir, x), data_transforms[x]) for x in ['train', 'valid']}
# Create training and validation dataloaders
dataloaders_dict = {x: torch.utils.data.DataLoader(image_datasets[x], batch_size=batch_size, shuffle=True, num_workers=4) for x in ['train', 'valid']}

# Detect if we have a GPU available
device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
print('device:', device)

image_datasets

# Send the model to GPU
model_ft = model_ft.to(device)

# Gather the parameters to be optimized/updated in this run. If we are
#  finetuning we will be updating all parameters. However, if we are
#  doing feature extract method, we will only update the parameters
#  that we have just initialized, i.e. the parameters with requires_grad
#  is True.
params_to_update = model_ft.parameters()
print("Params to learn:")
if feature_extract:
    params_to_update = []
    for name,param in model_ft.named_parameters():
        if param.requires_grad == True:
            params_to_update.append(param)
            print("\t",name)
else:
    for name,param in model_ft.named_parameters():
        if param.requires_grad == True:
            print("\t",name)

# Observe that all parameters are being optimized
optimizer_ft = optim.SGD(params_to_update, lr=1e-4, momentum=0.9)

def train_model(model, dataloaders, criterion, optimizer, num_epochs=25):
    since = time.time()

    val_acc_history = []
    
    best_model_wts = copy.deepcopy(model.state_dict())
    best_acc = 0.0
    
    for epoch in range(num_epochs):
        print('Epoch {}/{}'.format(epoch, num_epochs - 1))
        print('-' * 10)

        # Each epoch has a training and validation phase
        for phase in ['train', 'valid']:
            if phase == 'train':
                model.train()  # Set model to training mode
            else:
                model.eval()   # Set model to evaluate mode

            running_loss = 0.0
            running_corrects = 0

            # Iterate over data.
            for inputs, labels in tqdm(dataloaders[phase]):
                inputs = inputs.to(device)
                labels = labels.to(device)

                # zero the parameter gradients
                optimizer.zero_grad()

                # forward
                # track history if only in train
                with torch.set_grad_enabled(phase == 'train'):
                    outputs = model(inputs)
                    loss = criterion(outputs, labels)

                    _, preds = torch.max(outputs, 1)

                    # backward + optimize only if in training phase
                    if phase == 'train':
                        loss.backward()
                        optimizer.step()

                # statistics
                running_loss += loss.item() * inputs.size(0)
                running_corrects += torch.sum(preds == labels.data)

            epoch_loss = running_loss / len(dataloaders[phase].dataset)
            epoch_acc = running_corrects.double() / len(dataloaders[phase].dataset)

            print('{} Loss: {:.4f} Acc: {:.4f}'.format(phase, epoch_loss, epoch_acc))

            # deep copy the model
            if phase == 'valid' and epoch_acc > best_acc:
                best_acc = epoch_acc
                best_model_wts = copy.deepcopy(model.state_dict())
            if phase == 'valid':
                val_acc_history.append(epoch_acc)
                
        print()

    time_elapsed = time.time() - since
    print('Training complete in {:.0f}m {:.0f}s'.format(time_elapsed // 60, time_elapsed % 60))
    print('Best val Acc: {:4f}'.format(best_acc))

    # load best model weights
    model.load_state_dict(best_model_wts)
    return model, val_acc_history

###### Setup the loss fxn
criterion = nn.CrossEntropyLoss()

# Train and evaluate
model_ft, hist = train_model(model_ft, dataloaders_dict, criterion, optimizer_ft, num_epochs=num_epochs)

path = os.path.join(CNN_SAVE_DIR, 'trained_model.pt')
assert not os.path.isfile(path), print('A model is already saved at this location')
print(f'Saving model to {path}')
torch.save(model_ft, path)

# lol added this muna
####### wait idk where this came from or who wrote this here - kirfsten
device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
model_ft = torch.load(r"C:\Users\PSHSWVC-CID\Documents\estimating-co2-ph\scripts\data\models\trained_model.pt", map_location=torch.device('cpu')).to(device)

# you can run below if you want to see the final accuracy on nightlights over the train set
model_ft.eval()   # Set model to evaluate mode

criterion = nn.CrossEntropyLoss()
running_loss = 0.0
running_corrects = 0
total = 0


# Iterate over data.
for inputs, labels in tqdm(dataloaders_dict['train']):
    inputs = inputs.to(device)
    labels = labels.to(device)

    # forward
    # track history if only in train
    with torch.set_grad_enabled(False):
        outputs = model_ft(inputs)
        loss = criterion(outputs, labels)

        _, preds = torch.max(outputs, 1)

    # statistics
    running_loss += loss.item() * inputs.size(0)
    running_corrects += torch.sum(preds == labels.data)
    
    total += len(preds)
        
print(running_corrects.double()/total)
